# src/docs/swagger.yaml
definitions:
  dtos.CommentCreate:
    properties:
      message:
        maxLength: 300
        minLength: 5
        type: string
    required:
    - message
    type: object
  dtos.CommentResponse:
    properties:
      dislikes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      message:
        type: string
      tweet:
        $ref: '#/definitions/dtos.TweetResponse'
      tweet_id:
        type: integer
      user:
        $ref: '#/definitions/dtos.UserResponse'
      user_id:
        type: integer
    type: object
  dtos.CommentUpdate:
    properties:
      message:
        maxLength: 300
        minLength: 5
        type: string
    type: object
...
  dtos.TweetResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/dtos.CommentResponse'
        type: array
      dislikes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/dtos.UserResponse'
        type: array
      message:
        type: string
      title:
        type: string
      user:
        $ref: '#/definitions/dtos.UserResponse'
      user_id:
        type: integer
    type: object
...
  dtos.UserResponse:
    properties:
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      mobile_number:
        type: string
      username:
        type: string
    type: object
...
  models.Comment:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      dislikes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      enabled:
        type: boolean
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      message:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      tweet:
        $ref: '#/definitions/models.Tweet'
      tweet_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Tweet:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      dislikes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      enabled:
        type: boolean
      id:
        type: integer
      likes:
        items:
          $ref: '#/definitions/models.User'
        type: array
      message:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      title:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      comment_dislikes:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comment_likes:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      enabled:
        type: boolean
      firstname:
        type: string
      followers:
        items:
          $ref: '#/definitions/models.User'
        type: array
      followings:
        items:
          $ref: '#/definitions/models.User'
        type: array
      id:
        type: integer
      lastname:
        type: string
      mobile_number:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      password:
        type: string
      tweet_dislikes:
        items:
          $ref: '#/definitions/models.Tweet'
        type: array
      tweet_likes:
        items:
          $ref: '#/definitions/models.Tweet'
        type: array
      tweets:
        items:
          $ref: '#/definitions/models.Tweet'
        type: array
      username:
        type: string
    type: object
  responses.Response:
    properties:
      error:
        type: string
      message:
        type: string
      result: {}
      status:
        type: boolean
      statusCode:
        type: integer
      validationError:
        items:
          $ref: '#/definitions/validations.ValidationError'
        type: array
    type: object
  validations.ValidationError:
    properties:
      field:
        type: string
      param:
        type: string
      tag:
        type: string
    type: object
info:
  contact: {}
paths:
  /comment/delete:
    delete:
      description: delete comment
      parameters:
      - description: user's username
        in: query
        name: username
        required: true
        type: string
      - description: user's password
        in: query
        name: password
        required: true
        type: string
      - description: comment's id
        in: query
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Delete Comment
      tags:
      - Comment
  /comment/dislike:
    post:
      description: dislike comment
      parameters:
      - description: user's username
        in: query
        name: username
        required: true
        type: string
      - description: user's password
        in: query
        name: password
        required: true
        type: string
      - description: comment's id
        in: query
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Dislike Comment
      tags:
      - Comment
  /comment/get/comment:
    get:
      description: get comment
      parameters:
      - description: comment's id
        in: query
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  $ref: '#/definitions/models.Comment'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get Comment
      tags:
      - Comment
  /comment/get/comments:
    get:
      description: get comments of an user
      parameters:
      - description: user's username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/dtos.CommentResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get Comments
      tags:
      - Comment
  /comment/like:
    post:
      description: like comment
      parameters:
      - description: user's username
        in: query
        name: username
        required: true
        type: string
      - description: user's password
        in: query
        name: password
        required: true
        type: string
      - description: comment's id
        in: query
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Like Comment
      tags:
      - Comment
  /comment/post:
    post:
      consumes:
      - application/json
      description: post comment with message
      parameters:
      - description: user's username
        in: query
        name: username
        required: true
        type: string
      - description: user's password
        in: query
        name: password
        required: true
        type: string
      - description: tweet's id
        in: query
        name: tweet_id
        required: true
        type: string
      - description: get message of new comment
        in: body
        name: CommentCreate
        required: true
        schema:
          $ref: '#/definitions/dtos.CommentCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  $ref: '#/definitions/dtos.CommentResponse'
              type: object
        "400":
          description: Validation Error
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Post Comment
      tags:
      - Comment
  /comment/update:
    put:
      consumes:
      - application/json
      description: update comment's message
      parameters:
      - description: user's username
        in: query
        name: username
        required: true
        type: string
      - description: user's password
        in: query
        name: password
        required: true
        type: string
      - description: comment's id
        in: query
        name: comment_id
        required: true
        type: string
      - description: get updated message of comment
        in: body
        name: CommentUpdate
        required: true
        schema:
          $ref: '#/definitions/dtos.CommentUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  $ref: '#/definitions/dtos.CommentResponse'
              type: object
        "400":
          description: Validation Error
          schema:
            $ref: '#/definitions/responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Update Comment
      tags:
      - Comment
  /file/post:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: create new file
      parameters:
      - in: formData
        name: description
        required: true
        type: string
      - description: create file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/responses.Response'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/dtos.UserResponse'
                  type: array
              type: object
        "406":
          description: Error
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Create File
      tags:
      - File
...
swagger: "2.0"